+++ docs/CHANGELOG.md
@@ -574,11 +574,27 @@
     <br>
     - Renamed email file from a min.js to just a js.
     </details>
+    <br>
+</details>
+
+<details>
+<summary>5.1</summary>
+    <br>
     <details>
     <summary>5.1.0 - (1/9/24 5:05pm)</summary>
     - Added Verification Modal
     <br>
     - Implemented Verification Code after signing up.
+    </details>
+    <br>
+</details>
+
+<details>
+<summary>5.2</summary>
+    <br>
+    <details>
+    <summary>5.2.0 - (1/9/24 6:25pm)</summary>
+    - Implemented Login Feature.
     </details>
     <br>
 </details>